Move Publish and Status panel to wp-story-editor
