+++ packages/wp-story-editor/package.json
@@ -22,12 +22,14 @@
     "@wordpress/api-fetch": "^5.2.4",
     "flagged": "^2.0.1",
     "polished": "^4.1.3",
+    "react-calendar": "^3.5.0",
     "styled-components": "^5.3.3"
   },
   "devDependencies": {
     "@storybook/addon-actions": "^6.3.6",
     "@testing-library/react": "^12.1.2",
     "@testing-library/react-hooks": "^7.0.1",
-    "prop-types": "^15.7.2"
+    "prop-types": "^15.7.2",
+    "mockdate": "^3.0.5"
   }
 }
