+++ packages/wp-story-editor/src/testUtils/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export { default as ExcerptPanel } from './excerpt';
-export { default as PageAdvancementPanel } from './pageAdvancement';
-export { default as SlugPanel } from './slug';
-export { default as BackgroundAudioPanel } from './backgroundAudio';
-export { default as TaxonomiesPanel } from './taxonomies';
+export { default as renderWithTheme } from './renderWithTheme';
